third_party/WebKit/LayoutTests/http/tests/websocket/cookie-http-to-ws-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check http://xhr.spec.whatwg.org/.
Test How WebSocket handles cookies with cookie-av's.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS cookie is "WK-websocket-test-domain=1; WK-websocket-test-path=1"
A WebSocket connection to ws://127.0.0.1:8880/echo-cookie is opened.
A WebSocket connection to ws://127.0.0.1:8880/echo-cookie is closed.
PASS cookie is "WK-websocket-test-domain=1"
PASS successfullyParsed is true

TEST COMPLETE
